feat: revamp && init add WASM frameworks support && add all MDN `input` attributes
