+++ examples/yew/Cargo.toml
@@ -1,13 +1,13 @@
 [package]
-name = "yew-components-tailwind"
-version = "0.1.1"
+name = "yew-input-rs"
+version = "0.1.0"
 authors = ["Mahmoud Harmouch <oss@wiseai.dev>"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-input_yew = { path = "../.." }
+input-rs = { path = "../..", features = ["yew"] }
 regex = { version = "1.9.1", default-features = false }
 reqwasm = { version = "0.5.0", default-features = false }
 serde = { version = "1.0.178", default-features = false }
