+++ examples/yew/assets/multi-step-form-one.png
