+++ examples/yew/index.html
@@ -8,6 +8,7 @@
       crossorigin="anonymous"
       defer
     ></script>
+    <link data-trunk rel="rust"/>
     <title>Yew Tailwind Forms</title>
     <base data-trunk-public-url />
   </head>
