+++ examples/yew/src/api.rs
