+++ examples/yew/src/api/auth.rs
