+++ examples/yew/src/app.rs
