+++ examples/yew/src/components.rs
