+++ examples/yew/src/components/common.rs
