+++ examples/yew/src/components/login_form_one.rs
@@ -1,5 +1,5 @@
 use crate::components::common::{validate_email, validate_input, LoginUserSchema};
-use input_yew::CustomInput;
+use input_rs::yew::Input;
 use regex::Regex;
 use wasm_bindgen_futures::spawn_local;
 use web_sys::{console, HtmlInputElement, Window};
@@ -18,21 +18,21 @@ pub fn login_form_one() -> Html {
     let password_valid_handle = use_state(|| true);
     let password_valid = (*password_valid_handle).clone();
 
-    let input_email_ref = use_node_ref();
-    let input_email_handle = use_state(|| "sad".to_string());
-    let input_email = (*input_email_handle).clone();
+    let email_ref = use_node_ref();
+    let email_handle = use_state(|| "sad".to_string());
+    let email = (*email_handle).clone();
 
-    let input_password_ref = use_node_ref();
-    let input_password_handle = use_state(|| "sad".to_string());
-    let input_password = (*input_password_handle).clone();
+    let password_ref = use_node_ref();
+    let password_handle = use_state(|| "sad".to_string());
+    let password = (*password_handle).clone();
 
     let onsubmit = Callback::from(move |event: SubmitEvent| {
         event.prevent_default();
 
-        let email_ref = input_password.clone();
-        let password_ref = input_password.clone();
+        let email_ref = password.clone();
+        let password_ref = password.clone();
         let error_handle = error_handle.clone();
-        console::log_1(&format!("Email: {}, Password: {}", input_email, input_password).into());
+        console::log_1(&format!("Email: {}, Password: {}", email, password).into());
 
         spawn_local(async move {
             let email_val = email_ref.clone();
@@ -67,36 +67,36 @@ pub fn login_form_one() -> Html {
                 <div class="mb-3 error bg-red-600 text-white px-4 py-3 font-semibold rounded-md text-center text-base">{error}</div>
               }
               <span class="text-4xl text-gray-800 pb-6">{"Login"}</span>
-              <CustomInput
-                input_type={"tek"}
+              <Input
+                r#type={"tek"}
                 label={"Email"}
-                input_handle={input_email_handle}
+                handle={email_handle}
                 name={"username"}
-                input_ref={input_email_ref}
-                input_placeholder={"Your Email"}
+                r#ref={email_ref}
+                placeholder={"Your Email"}
                 icon_class={"text-2xl fa fa-person text-gray-500 absolute left-3 top-1/2 transform -translate-y-1/2"}
                 error_message={"Enter a valid email address!"}
-                form_input_field_class={"relative"}
-                form_input_input_class={"input w-full px-4 py-3 rounded-lg border border-gray-300 bg-gray-100 pl-12"}
-                form_input_error_class={"text-red-500 absolute text-sm mb-2"}
+                field_class={"relative"}
+                input_class={"input w-full px-4 py-3 rounded-lg border border-gray-300 bg-gray-100 pl-12"}
+                error_class={"text-red-500 absolute text-sm mb-2"}
                 required={true}
-                input_valid_handle={email_valid_handle}
+                valid_handle={email_valid_handle}
                 validate_function={validate_email}
               />
-              <CustomInput
-                input_type={"password"}
+              <Input
+                r#type={"password"}
                 label={"Password"}
-                input_handle={input_password_handle}
+                handle={password_handle}
                 name={"password"}
-                input_ref={input_password_ref}
-                input_placeholder={"Password"}
+                r#ref={password_ref}
+                placeholder={"Password"}
                 icon_class={"text-2xl fa fa-lock text-gray-500 absolute left-3 top-1/2 transform -translate-y-1/2"}
                 error_message={"Password can't be blank!"}
-                form_input_field_class={"relative mt-2 mb-2"}
-                form_input_input_class={"input w-full px-4 py-3 rounded-lg border border-gray-300 bg-gray-100 pl-12"}
-                form_input_error_class={"text-red-500 absolute text-sm"}
+                field_class={"relative mt-2 mb-2"}
+                input_class={"input w-full px-4 py-3 rounded-lg border border-gray-300 bg-gray-100 pl-12"}
+                error_class={"text-red-500 absolute text-sm"}
                 required={true}
-                input_valid_handle={password_valid_handle}
+                valid_handle={password_valid_handle}
                 validate_function={validate_input}
                 eye_active={"cursor-pointer absolute right-4 top-1/2 transform -translate-y-1/2 text-2xl text-gray-600 toggle-button fa fa-eye"}
                 eye_disabled={"cursor-pointer absolute right-4 top-1/2 transform -translate-y-1/2 text-2xl text-gray-600 toggle-button fa fa-eye-slash"}
