+++ examples/yew/src/pages.rs
