+++ examples/yew/src/pages/contact_page_one.rs
@@ -4,7 +4,5 @@ use crate::components::contact_form_one::ContactFormOne;
 
 #[function_component(ContactPageOne)]
 pub fn contact_page_one() -> Html {
-    html! {
-        <ContactFormOne />
-    }
+    html! { <ContactFormOne /> }
 }
