+++ examples/yew/src/pages/error.rs
