+++ examples/yew/src/pages/login_page_one.rs
@@ -4,7 +4,5 @@ use crate::components::login_form_one::LoginFormOne;
 
 #[function_component(LoginPageOne)]
 pub fn login_page_one() -> Html {
-    html! {
-        <LoginFormOne />
-    }
+    html! { <LoginFormOne /> }
 }
