+++ examples/yew/src/pages/login_page_three.rs
@@ -4,7 +4,5 @@ use crate::components::login_form_three::LoginFormThree;
 
 #[function_component(LoginPageThree)]
 pub fn login_page_three() -> Html {
-    html! {
-        <LoginFormThree />
-    }
+    html! { <LoginFormThree /> }
 }
