+++ examples/yew/src/pages/login_page_two.rs
@@ -4,7 +4,5 @@ use crate::components::login_form_two::LoginFormTwo;
 
 #[function_component(LoginPageTwo)]
 pub fn login_page_two() -> Html {
-    html! {
-        <LoginFormTwo />
-    }
+    html! { <LoginFormTwo /> }
 }
