+++ examples/yew/src/pages/multi_step_page_one.rs
@@ -4,7 +4,5 @@ use crate::components::multi_step_form_one::MultiStepFormOne;
 
 #[function_component(MultiStepPageOne)]
 pub fn multi_step_page_one() -> Html {
-    html! {
-        <MultiStepFormOne />
-    }
+    html! { <MultiStepFormOne /> }
 }
