+++ examples/yew/src/router.rs
