+++ examples/yew/tailwind.config.js
